push* scripts: use luasrcdiet if directed
If ${LUADIET} is empty, send files as they are; otherwise, stream them
through luasrcdiet.  To use the default options, use 'LUADIET=' '" or
something to that effect.  This can be set per-file, should some module
turn out to be broken by dieting.

The "dopush" utility macro is split and now called "dopushlua" and
"dopushtext" (which skips luasrcdiet even if it is engaged).