Alternating score display
Implements Tim's suggestions in https://github.com/cmukgb/ctfws-timer-iot/issues/5

If it all fits on one line, do that.  If it's absurdly long, scroll the
whole thing.  In the middle, switch between displaying the two scores
(and always displaying the number of flags), padded so that the
rightmost digits align.